Misc fixes/changes
- Move user search into User class
- Inline permissions assignment in workspace create
- Fix Workspace permissions use
- Remove staggered use of Workspace().permissions, and get_permissions.
- Add readonly flag to uploads_database
- Remove unnecessary cast to bool in `is_reader`
- Remove extra call when getting table rows in csv downloader
- Narrow type of arango Graph.vertex stub
- Add extra elif to node_edges func
- Remove immediate flag in db.db()
