+++ mypy_stubs/arango/graph.pyi
@@ -1,4 +1,4 @@
-from typing import Dict, List, Optional, Any
+from typing import Dict, List, Optional, Union, Any
 from arango.collection import EdgeCollection, VertexCollection  # type: ignore
 
 class Graph:
@@ -7,5 +7,5 @@ class Graph:
     def vertex_collections(self) -> List[str]: ...
     def edge_definitions(self) -> List[Dict]: ...
     def vertex(
-        self, vertex: Any, rev: Optional[Any] = ..., check_rev: bool = ...
+        self, vertex: Union[Dict, str], rev: Optional[Any] = ..., check_rev: bool = ...
     ) -> Dict: ...
